Update polymorphism.md (#371)
When I initially read this, I thought that specifying the 'delegate' kind was the way to opt in to the feature. If someone is as careless as I was, they might not realize they are turning off access policies by only specifying one kind. I think this change will help the reader realize that each kind has to be specified